take apart gameSettings into 2 modules
- game settings really had 2 different logic changes. Handling board size settings and opponent type settings
